pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rlm-repo-intel"
version = "0.1.0"
description = "RLM-powered repository intelligence — analyze codebases, PRs, and issues using recursive language models"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [
    { name = "Vishal Tandale", email = "vishal@relayerlabs.ai" },
]

dependencies = [
    "rlms>=0.1.0",
    "click>=8.0",
    "httpx>=0.27",
    "psycopg2-binary>=2.9",
    "pyyaml>=6.0",
    "rich>=13.0",
    "tree-sitter>=0.23",
    "tiktoken>=0.7",
    "numpy>=1.26",
    "litellm>=1.81.13",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.5",
]

[project.scripts]
rlm-repo-intel = "rlm_repo_intel.cli:main"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
